Rename predicates -> queries
