+++ src/world/Tiles/Arcology.ts
@@ -7,8 +7,8 @@ import Game from "../../Game.js";
 import { AiResearchTech, RationalityTech, CognitiveBiasesTech } from "../../techtree/TechTree.js";
 import TileProject from "../../tileProjects/TileProject.js";
 import { stripIndent } from "../../util/Text.js";
-import { techRequirement } from "../../predicates/DescribedTileQuery.js";
-import { hasTech, speciesHasPopulation, notQuery } from "../../predicates/predicates.js";
+import { techRequirement } from "../../queries/DescribedTileQuery.js";
+import { hasTech, speciesHasPopulation, notQuery } from "../../queries/Queries.js";
 import { Schemas as S } from "../../serialize/Schema.js";
 
 export default class Arcology extends Tile {
