+++ src/world/Tiles/ConstructionHabitat.ts
@@ -12,8 +12,8 @@ import Arcology from "./Arcology.js";
 import Wasteland from "./Wasteland.js";
 import { constructionProject } from "../../tileProjects/TileProject.js";
 import Hydroponics from "./Hydroponics.js";
-import { techRequirement, roadRequirement } from "../../predicates/DescribedTileQuery.js";
-import { hasTech } from "../../predicates/predicates.js";
+import { techRequirement, roadRequirement } from "../../queries/DescribedTileQuery.js";
+import { hasTech } from "../../queries/Queries.js";
 import RobotHive from "./RobotHive.js";
 import { Schemas as S } from "../../serialize/Schema.js";
