+++ src/world/Tiles/Monolith.ts
@@ -13,8 +13,8 @@ import Resource from "../../resources/Resource.js";
 import Cost from "../../resources/Cost.js";
 import Game from "../../Game.js";
 import HumanMonolith from "./HumanMonolith.js";
-import { tileWithinDistanceRequirement, techRequirement } from "../../predicates/DescribedTileQuery.js";
-import { hasTech } from "../../predicates/predicates.js";
+import { tileWithinDistanceRequirement, techRequirement } from "../../queries/DescribedTileQuery.js";
+import { hasTech } from "../../queries/Queries.js";
 import { Schemas as S } from "../../serialize/Schema.js";
 
 export default class Monolith extends Tile {
