+++ src/world/Tiles/Wasteland.ts
@@ -14,8 +14,8 @@ import { MonolithSurveyTech } from "../../techtree/TechTree.js";
 import ConstructionVictory from "./ConstructionVictory.js";
 import NeuralEmulator from "./NeuralEmulator.js";
 import NanotechFoundry from "./NanotechFoundry.js";
-import { hasTech, tileExists, orQuery, notQuery } from "../../predicates/predicates.js";
-import { roadRequirement } from "../../predicates/DescribedTileQuery.js";
+import { hasTech, tileExists, orQuery, notQuery } from "../../queries/Queries.js";
+import { roadRequirement } from "../../queries/DescribedTileQuery.js";
 import { Schemas as S } from "../../serialize/Schema.js";
 
 export default class Wasteland extends Tile {
